Persist/restore launcher window geometry and improve display/resolution handling
### Motivation

- Replace ad-hoc QSettings usage with persisted `settings` schema for launcher main window geometry so window position/size survive restarts and multiple-platform setups.
- Ensure the launcher window remains visible when displays are hot-plugged or the primary screen changes.
- Improve display/resolution selection and scaling logic to avoid oversized/stale resolutions and to respect selected display index.

### Description

- Extended `config/schemas/settings.json` to add `launcher.mainWindow.geometry` with fields `valid`, `x`, `y`, `width`, and `height`.
- Replaced `QSettings` usage with reading/writing `settings["launcher"]["mainWindow"]["geometry"]` and added `saveWindowSettings`, `saveWindowSettingsUnchecked`, and `ensureWindowVisibleOnExistingScreen` to `MainWindow` to persist and validate geometry.
- Added logic to relocate/center the launcher on screen changes by connecting to `QGuiApplication::screenRemoved` and `primaryScreenChanged` and scheduling relocation retries.
- Removed `CLauncherDirs::getSettings` dependency and updated header with a forward declaration of `QScreen`.
- In `CSettingsView` added handling for `video.displayIndex` bounds, used SDL to query desktop resolution (`findDesktopResolution`) and ensured windowed resolutions are only kept if they fit the selected display; updated resolution fallback logic to persist a safe replacement resolution when needed.
- Minor fixes: small SDL loop whitespace fix and adjusted scaling range refresh when display index changes.

### Testing

- Performed an automated desktop build of the launcher with the changes and the build completed successfully.
- Ran the project's automated unit/integration test suite and it passed without regressions.

+++ launcher/settingsView/csettingsview_moc.cpp
@@ -91,7 +91,7 @@ void CSettingsView::setDisplayList()
 	{
 		displayIndex = 0;
 		Settings node = settings.write["video"]["displayIndex"];
-		node->Float() = displayIndex;
+		node->Integer() = displayIndex;
 	}
 
 	if(list.count() < 2)
@@ -548,8 +548,8 @@ void CSettingsView::on_comboBoxResolution_currentTextChanged(const QString & arg
 	QStringList list = arg1.split("x");
 
 	Settings node = settings.write["video"]["resolution"];
-	node["width"].Float() = list[0].toInt();
-	node["height"].Float() = list[1].toInt();
+	node["width"].Integer() = list[0].toInt();
+	node["height"].Integer() = list[1].toInt();
 
 	fillValidResolutions();
 	fillValidScalingRange();
@@ -583,7 +583,7 @@ void CSettingsView::on_buttonFullModExtraction_toggled(bool value)
 void CSettingsView::on_comboBoxDisplayIndex_currentIndexChanged(int index)
 {
 	Settings node = settings.write["video"];
-	node["displayIndex"].Float() = index;
+	node["displayIndex"].Integer() = index;
 
 	fillValidResolutionsForScreen(index);
 	fillValidScalingRange();
